TINY-9157: Move TestStore out of alloy and into agar
