+++ modules/agar/src/main/ts/ephox/agar/api/TestStore.ts
