+++ modules/alloy/src/main/ts/ephox/alloy/api/testhelpers/TestHelpers.ts
@@ -1,8 +1,6 @@
 // Test code - should eventually move to a separate project
 import * as GuiSetup from './GuiSetup';
-import { TestStore } from './TestStore';
 
 export {
-  GuiSetup,
-  TestStore
+  GuiSetup
 };
