+++ modules/tinymce/src/themes/silver/test/ts/headless/window/CustomDialogTest.ts
@@ -1,4 +1,4 @@
-import { FocusTools, Keyboard, Keys, Mouse, UiFinder, Waiter } from '@ephox/agar';
+import { FocusTools, Keyboard, Keys, Mouse, TestStore, UiFinder, Waiter } from '@ephox/agar';
 import { TestHelpers } from '@ephox/alloy';
 import { before, describe, it } from '@ephox/bedrock-client';
 import { Arr, Fun } from '@ephox/katamari';
@@ -13,7 +13,7 @@ import * as TestExtras from '../../module/TestExtras';
 const GuiSetup = TestHelpers.GuiSetup;
 
 describe('headless.tinymce.themes.silver.window.CustomDialogTest', () => {
-  const store = TestHelpers.TestStore();
+  const store = TestStore();
   const helpers = TestExtras.bddSetup();
   let windowManager: WindowManagerImpl;
   before(() => {
