+++ modules/tinymce/src/themes/silver/test/ts/headless/window/SilverUrlDialogTest.ts
@@ -1,5 +1,4 @@
-import { ApproxStructure, Assertions, Mouse, UiFinder, Waiter } from '@ephox/agar';
-import { TestHelpers } from '@ephox/alloy';
+import { ApproxStructure, Assertions, Mouse, TestStore, UiFinder, Waiter } from '@ephox/agar';
 import { before, describe, it } from '@ephox/bedrock-client';
 import { SugarBody } from '@ephox/sugar';
 
@@ -10,7 +9,7 @@ import * as WindowManager from 'tinymce/themes/silver/ui/dialog/WindowManager';
 import * as TestExtras from '../../module/TestExtras';
 
 describe('headless.tinymce.themes.silver.window.SilverUrlDialogTest', () => {
-  const store = TestHelpers.TestStore();
+  const store = TestStore();
   const helpers = TestExtras.bddSetup();
   let windowManager: WindowManagerImpl;
   let dialogApi: Dialog.UrlDialogInstanceApi;
