+++ versions.txt
@@ -1,3 +1,5 @@
 # List of packages to bump:
 # Format: [package_name]@[new_version]
 
+alloy@12.0.0
+agar@7.3.0
